pyproject.toml
[project]
name = "ev-price-tracker-backend"
version = "0.0.1"
description = "EV Price Tracker backend to handle data requests from the frontend"
authors = [{ name = "Issac Loo", email = "issaloo@umich.edu" }]

requires-python = ">=3.10,<3.11"
license = { text = "All Rights Reserved" }
dependencies = [
    "psycopg2>=2.9.9",
    "google-cloud>=0.34.0",
    "google-cloud-secret-manager>=2.16.4",
    "python-dotenv>=1.0.0",
    "django>=4.2.6",
    "djangorestframework>=3.14.0",
    "django-redis>=5.4.0",
    "redis>=5.0.1",
    "djangorestframework-api-key>=3.0.0",
    "djangorestframework-simplejwt>=5.3.1",
    "bcrypt>=4.1.2",
    "django-cors-headers>=4.3.1",
]

[project.optional-dependencies]
dev = [
    "cookiecutter>=2.1.1",
    "pytest>=7.3.1",
    "commitizen>=3.6.0",
    "pre-commit>=3.3.3",
    "gitlint>=0.19.1",
    "scalene>=1.5.26",
    "pandera>=0.16.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
select = [
    "B",  # flake8-bugbear
    "C",  # mccabe complexity
    "E",  # Error
    "F",  # pyflakes
    "W",  # Warnings
    "I",  # isort
    "PD", # pandas-vet https://beta.ruff.rs/docs/rules/#pandas-vet-pd
    "D",  # pydocstyle https://beta.ruff.rs/docs/rules/#pydocstyle-d
]
ignore = [
    "B008", # function-call-in-default-argument
    "B905", # zip-without-explicit-strict
    "C901", # complex-structure
    "D100", # undocumented-public-module
    "D104", # undocumented-public-package
    "D205", # blank-line-after-summary
    "D211", # blank-line-before-class
    "D212", # multi-line-summary-first-line
    "D213", # multi-line-summary-second-line
    "D417", # undocumented-param
    "E501", # line-too-long (handled by black)
    "F403", # unable-to-detect-undefined-import-names
]

[tool.black]
line-length = 120
